% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{college_load_tidy}
\alias{college_load_tidy}
\alias{college_load_tidy_all}
\alias{college_load_tidy_scorecard}
\alias{college_load_tidy_school}
\title{Load the Tidied College Scorecard Data Set}
\usage{
college_load_tidy_all()

college_load_tidy_scorecard()

college_load_tidy_school()
}
\value{
A list with two data frames: \code{scorecard} and \code{school}.
}
\description{
Loads a tidied version of the U.S. College Scorecard data set, either
individually or as a list with two data frames: \code{scorecard} and \code{school}.
The data set contain only colleges that participate in Title IV federal
financial aid programs. In the tidy data set, column names have been changed
for readability and consistency.

You can find the original data set at the U.S. Department of Education's
\href{https://collegescorecard.ed.gov/data/}{College Scorecard website}. The
code for the data transformation process is available at
\url{https://github.com/gadenbuie/scorecard-db}.
}
\section{Functions}{
\itemize{
\item \code{college_load_tidy_all()}: Load both the tidied scorecard and school data
sets.

\item \code{college_load_tidy_scorecard()}: Load the tidy scorecard data set.

\item \code{college_load_tidy_school()}: Load the tidy school data set.

}}
\examples{
c(scorecard, school) \%<-\% college_load_tidy_all()
scorecard <- college_load_tidy_scorecard()
school <- college_load_tidy_school()

}
