% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collegeScorecard-package.R
\docType{package}
\name{collegeScorecard-package}
\alias{collegeScorecard}
\alias{collegeScorecard-package}
\title{collegeScorecard: US College Scorecard Data}
\description{
Provides a tidied subset of the US College Scorecard dataset, containing institutional characteristics, enrollment, student aid, costs, and student outcomes at institutions of higher education in the United States.
}
\section{Updated}{


The data were last updated on \strong{April 23, 2025} from
\url{https://collegescorecard.ed.gov/data/}. Please visit
\url{https://github.com/gadenbuie/scorecard-db} for the code used to process
and transform the released data into the tidy data sets available in this
package.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gadenbuie/scorecard-db}
  \item \url{https://collegescorecard.ed.gov/data/}
  \item Report bugs at \url{https://github.com/gadenbuie/scorecard-db/issues}
}

}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{garrick@adenbuie.com} (\href{https://orcid.org/0000-0002-7111-0077}{ORCID}) [copyright holder]

}
\keyword{internal}
