% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{cs_load_tidy}
\alias{cs_load_tidy}
\title{Load the Tidied College Scorecard Data Set}
\usage{
cs_load_tidy()
}
\value{
A list with two data frames: \code{scorecard} and \code{school}.
}
\description{
Loads a tidied version of the College Scorecard data set, returning
a list with two data frames: \code{scorecard} and \code{school}.
}
\examples{
c(scorecard, school) \%<-\% cs_load_tidy()

}
