% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collegeScorecard-package.R
\docType{package}
\name{collegeScorecard-package}
\alias{collegeScorecard}
\alias{collegeScorecard-package}
\title{collegeScorecard: U.S. College Scorecard Data}
\description{
The US College Scorecard dataset provides information on institutional characteristics, enrollment, student aid, costs, and student outcomes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gadenbuie/scorecard-db}
  \item Report bugs at \url{https://github.com/gadenbuie/scorecard-db/issues}
}

}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{garrick@adenbuie.com} (\href{https://orcid.org/0000-0002-7111-0077}{ORCID})

}
\keyword{internal}
